input: touchscreen: imx6ul_tsc: convert int to u32
The code uses of_property_read_u32 and expects positive values.
However, the values are stored in signed int variables.
Additionally, the registers values are also stored in signed
variables without a good reason (readl/writel expect u32).

The only time this caused a real bug was in the new average-samples
property, in which the property is numerically compared and
implicitly expected to be positive.
I believe it's better to change all the properties and registers to
u32, for consistency and warnings reduction.

Signed-off-by: Guy Shapiro <guy.shapiro@mobi-wize.com>
Reported-by: Bjørn Forsman <bjorn.forsman@gmail.com>